Fixed some tests
git-svn-id: https://svn.origo.ethz.ch/zoodb/trunk@300 b278144e-e56a-4748-98cb-974f122dab01

+++ src/org/zoodb/jdo/internal/model1p/Node1P.java
@@ -116,7 +116,10 @@ public class Node1P extends Node {
 		//Deleting objects class-wise reduces schema index look-ups (negligible?) and allows batching. 
 		for (Entry<ZooClassDef, ArrayList<PersistenceCapableImpl>> entry: toDelete.entrySet()) {
 			ZooClassDef clsDef = entry.getKey();
-			disk.deleteObjects(clsDef.getOid(), entry.getValue());
+			//Ignore instances of deleted classes, there is a dropInstances for them
+			if (!clsDef.jdoIsDeleted() && !commonCache.getCachedSchema(clsDef.getJavaClass(), this).isDeleted()) {
+				disk.deleteObjects(clsDef.getOid(), entry.getValue());
+			}
 			entry.setValue(null);
 		}
